Animations update
Final animations placed in, including new run, idle, jump, fall and push sprites.
New sprite sheets are placed in the the "Player Sprites" Folder. Push_F7 is seperate due to the one on the sprite sheet being the wrong size. Also added a push animation to the player. it just needs to be connected via code.